fix checkbox state behaviour
