+++ src/components/checkbox/Checkbox.stories.mdx
@@ -1,5 +1,6 @@
 import {Canvas, Meta, Story, ArgsTable, Source} from "@storybook/addon-docs/blocks";
-import Checkbox, {LABEL_PLACEMENT} from "./Checkbox"
+import Checkbox, {LABEL_PLACEMENT} from "./Checkbox";
+import CheckboxStoryWidget from "./ui/CheckboxStoryWidget";
 
 <Meta
   title="Field/Checkbox"
@@ -8,7 +9,7 @@ import Checkbox, {LABEL_PLACEMENT} from "./Checkbox"
 
 export const Template = ({...args}) => {
   return (
-    <Checkbox {...args} />
+    <CheckboxStoryWidget {...args} />
   );
 }
